fix : 파일 경로 변경
- MainPage -> Main
